Added the "initial" field to ObservableValueEvent.  This will allow observers to determine whether an event represents the initial value poplation event of an ObservableValue or a change in value from a null value to something else.  Currently this is not possible.  In particular, this inability could cause problems with collections containing null elements.
I have fixed all the compile errors, but I have not gone to all the call sites of ObservableValue.createChangeEvent() (renamed) to correct them to createInitialEvent() (new) where applicable.  I also need to go to all the call sites of getOldValue() and convert most of the null checks to a test on the "initial" flag.

So there may still be problems with null values in observables, but it's a start, and things should work no worse than they did before.  All tests pass except the new one that doesn't work yet.